tui: Move CLI command escape
Interactive mode now delegates shell escapes, scoped menus, fixup
selection, and hunk action routing to focused TUI modules.

The fallback CLI command path still keeps shlex parsing and non-interactive
command execution inside the dispatcher.

This commit addresses that by moving the fallback CLI command path into
tui.cli_escape and passing the help renderer from _dispatch_action.

Subsequent commits will retarget the CLI escape guard and add owner-level
tests.
